feat(add): add stock and sizes display with description for items
